испправил логику в соответсвиеем с указанием в ревью
