+++ src/service/TaskManager.java
@@ -6,13 +6,12 @@ import model.SubTask;
 import model.Task;
 
 import java.util.List;
-import java.util.TreeSet;
 
 public interface TaskManager {
 
     List<Task> getHistory();
 
-    TreeSet<Task> getPrioritizedTasks();
+    List<Task> getPrioritizedTasks();
 
     Task create(Task task) throws ManagerSaveException;
